requirements.txt
requests
djangorestframework
pyTelegramBotAPI
python-dotenv
django-filter